ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
aiohttp==3.9.5
sentence-transformers==2.7.0
groq==0.5.0
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1
pydantic==2.7.0
lxml==5.2.1
